Add random password feature
We need to manage user's channels and other resources using
Terraform without knowledge of user's actual password (due to
privacy issues).

So the following feature is implemented.

If password field in user resource definition is set to "-", the
user is created with a random password and "reset password" email
is sent to the user.